Search all reachable commits for upstream branch commit
Search the full history of all commits in the reflog of the upstream branch to find the first common commit in the child branch, using a window of visibility based on the commit timestamp to keep memory footprint reasonable, without relying too heavily on not having clock skew.

Unfortunately, this does not appear to match what git itself does (e.g. in git pull); it appears to only consider the log of the current commit, and the reflog commits of the branch, not the log of the reflog commits.